+++ src/veins/modules/mobility/traci/TraCIScenarioManager.cc
@@ -263,7 +263,7 @@ void TraCIScenarioManager::initialize(int stage) {
 
 	vehicleNameCounter = 0;
 	vehicleRngIndex = par("vehicleRngIndex");
-	numVehicles = par("numVehicles").longValue();
+	numVehicles = par("numVehicles");
 	mobRng = getRNG(vehicleRngIndex);
 
 	annotations = AnnotationManagerAccess().getIfExists();
